Create SwapCase.js
Have the function SwapCase(str) take the str parameter and swap the case of each character. For example: if str is "Hello World" the output should be hELLO wORLD. Let numbers and symbols stay the way they are. 

Use the Parameter Testing feature in the box below to test your code with different arguments.